server/requirements.txt
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
